refactor: split simulation runner phase executors
